test(transactions): apply waitForStableTxList helper to edit-layout test
The same loading-state race that was previously fixed in
transactions-promote-recurring.test.tsx (PR #197 commit 502f771) was
still latent in transactions-edit-layout-and-settled-date.test.tsx.
The findAllByText + waitFor(queryAllByRole(/Edit:/)) combo can land
between the spinner re-render and the post-loadRefs Edit-button mount,
intermittently dropping the just-set editingId on the Edit click.

Extract waitForStableTxList to frontend/tests/utils/ as the shared
source of truth, import it from both test files, and apply it to all
seven racy spots in the edit-layout/settled-date test.

Verified locally: edit-layout 12/12, promote-recurring 8/8, full suite
323/323, tsc clean, 5x repeat run on both racy files all green.
